kvs: cache all intention nodes; track touched nodes
this patch (1) caches all nodes in an intention on a cache miss, rather
than only the node that was requested. it also (2) tracks an estimate of
how many nodes in the cache are touched (i.e. have been references since
they were added to the cache) which helps give us an idea about how much
work reading nodes is wasted.

the stat is an estimate because the (1) the counters are sloppy and (2)
there is a lot of places where nodes are referenced for various reasons
and a lot of indirection in how nodes are counted. so it may be that we
aren't hitting every spot, but it should be close enough for getting
started.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>